refactor: extract pr-reviewer-modal as separate component
